v3.6
Very minor change. The idea of functions is that they are self-contained and do not affect any hidden state. strtosize_t() in primitives.c was failing this promise. It expected errno to have been 0 when it was called, and left it unchanged if it was altered by strtoumax(). Fixed this - it now saves the old value, sets errno to 0, and then restores the old value after calling strtoumax().